+++ rsenv/origami/fileconverters/cadnano_seqpool_csv_to_idt_xls_orderfile.py
@@ -1,0 +1,9 @@
+# Copyright 2019, Rasmus Sorensen <rasmusscholer@gmail.com>
+"""
+
+Module for creating an IDT Excel "order file".
+
+See also the old `oligomanager_tools.file_transformation.design_ordering` module.
+
+
+"""